Return back performance in multiproject
