+++ backend.native/tests/build.gradle
@@ -3098,7 +3098,7 @@ kotlinNativeInterop {
         defFile 'interop/basics/ccallbacksAndVarargs.def'
     }
 
-    if (isAppleTarget(project)) {
+    if (PlatformInfo.isAppleTarget(project)) {
         objcSmoke {
             defFile 'interop/objc/objcSmoke.def'
             headers "$projectDir/interop/objc/smoke.h"
@@ -3447,7 +3447,7 @@ task runKonanRegexTests(type: RunKonanTest) {
 
 task coverage_basic_program(type: RunStandaloneKonanTest) {
 
-    disabled = getTarget(project).name != "ios_x64" && getTarget(project).name != "macos_x64"
+    disabled = PlatformInfo.isAppleTarget(project)
 
     description = "Test that `-Xcoverage` generates correct __llvm_coverage information"
 
@@ -3476,7 +3476,7 @@ task coverage_basic_program(type: RunStandaloneKonanTest) {
 
 task coverage_basic_library(type: RunStandaloneKonanTest) {
 
-    disabled = getTarget(project).name != "ios_x64" && getTarget(project).name != "macos_x64"
+    disabled = PlatformInfo.isAppleTarget(project)
 
     description = "Test that `-Xlibrary-to-cover` generates correct __llvm_coverage information"
