+++ buildSrc/plugins/src/main/kotlin/org/jetbrains/kotlin/RunKotlinNativeTask.kt
@@ -3,6 +3,8 @@
  * that can be found in the license/LICENSE.txt file.
  */
 
+package org.jetbrains.kotlin
+
 import groovy.lang.Closure
 import org.gradle.api.DefaultTask
 import org.gradle.api.Task
