--- performance/buildSrc/build.gradle
@@ -1,36 +1,0 @@
-buildscript {
-    repositories {
-        maven {
-            url 'https://cache-redirector.jetbrains.com/maven-central'
-        }
-        maven {
-            url kotlinCompilerRepo
-        }
-        jcenter()
-    }
-
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
-    }
-}
-
-apply plugin: 'kotlin'
-
-repositories {
-    maven {
-        url kotlinCompilerRepo
-    }
-    maven {
-        url buildKotlinCompilerRepo
-    }
-    jcenter()
-}
-
-dependencies {
-    compileOnly gradleApi()
-    implementation "org.jetbrains.kotlin:kotlin-gradle-plugin"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-common:$kotlinVersion"
-    compile group: 'com.ullink.slack', name: 'simpleslackapi', version: '1.2.0'
-}
-
-sourceSets.main.kotlin.srcDirs = ["src", "$projectDir/../../tools/benchmarks/shared/src"]
