+++ performance/cinterop/build.gradle
@@ -5,7 +5,7 @@ project.ext {
     nativeSrcDirs = ['src/main/kotlin-native', '../shared/src/main/kotlin-native']
 
 }
-apply from: rootProject.file("${rootProject.rootDir}/gradle/benchmark.gradle")
+apply from: rootProject.file("${rootProject.rootDir}/performance/gradle/benchmark.gradle")
 kotlin.targets.native.compilations.main.cinterops { 
     macros
     struct
