+++ performance/helloworld/build.gradle
@@ -7,4 +7,4 @@ project.ext {
     buildSteps = ["runKonanc": ["$dist/bin/konanc$toolExtension", "${project.getProjectDir()}/src/main/kotlin/main.kt", "-o",
                                 "${buildDir.absolutePath}/program${MPPTools.getNativeProgramExtension()}"]]
 }
-apply from: rootProject.file("${rootProject.rootDir}/gradle/compileBenchmark.gradle")
+apply from: rootProject.file("${rootProject.rootDir}/performance/gradle/compileBenchmark.gradle")
