+++ performance/objcinterop/build.gradle
@@ -4,7 +4,7 @@ project.ext {
     jvmSrcDirs = ['src/main/kotlin-jvm', '../shared/src/main/kotlin-jvm']
     nativeSrcDirs = ['src/main/kotlin-native', '../shared/src/main/kotlin-native']
 }
-apply from: rootProject.file("${rootProject.rootDir}/gradle/benchmark.gradle")
+apply from: rootProject.file("${rootProject.rootDir}/performance/gradle/benchmark.gradle")
 kotlin.targets.native.compilations.main.cinterops {
     classes {
         headers "$projectDir/src/nativeInterop/cinterop/complexNumbers.h"
