--- performance/settings.gradle
@@ -1,6 +1,0 @@
-include ':ring'
-include ':cinterop'
-include ':helloworld'
-include ':videoplayer'
-include ':framework'
-include ':objcinterop'
